requirements.txt
# Core dependencies
torch
transformers>=4.35.0
modelscope==1.9.5
huggingface_hub>=0.21.4
accelerate>=0.26.1

# Audio/text processing (used in original main.py)
langchain>=0.1.14
langchain_groq>=0.1.4
python-dotenv>=1.0.1

# Additional dependencies used by ModelScope pipelines
opencv-python
imageio
imageio-ffmpeg
numpy
s3fs
aiobotocore==2.5.4
botocore>=1.31.17,<1.31.18

# Optional: for progress bar and utility enhancements
tqdm